projects/dl/llm/llm-med-project/pyproject.toml
[project]
name = "llm-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "bert-score>=0.3.13",
    "bitsandbytes>=0.49.0",
    "datasets>=2.20.0",
    "evaluate>=0.4.6",
    "fsspec>=2024.3.1",
    "huggingface-hub>=0.36.0",
    "hydra-core>=1.3.2",
    "ipykernel>=7.1.0",
    "peft>=0.18.0",
    "pyarrow>=22.0.0",
    "pytorch-lightning>=2.6.0",
    "rouge-score>=0.1.2",
    "scikit-learn>=1.8.0",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.3",
    "wandb>=0.23.1",
]
